Rename package 'eval' to 'px' and change how runtime is accessed.
This commit perfoms some refactoring motivated by the fact that nothing
in this module is related to evaluation. The following refactoring is
done:

* Package 'eval' is renamed to 'px'
* Packages 'pcore' and 'impl' are merged into a 'pximpl' package
* A new 'pcore' package is added that contains global functions to
  access the pcore runtime instance. This means that expressions like:

    eval.Puppet.Do(...)

must be replaced with:

    pcore.Do(...)

This also means that there's no longer a need to have underscore imports
to ensure the initialization of the pcore runtime.

Relates to lyraproj/lyra#137