Fix problem with reflecting nested maps and slices into interface{}
A type declared as nested slice or map where the leaf type was an
interface{} would fail assignment of slices or maps to that leaf. The
reason was that the type of the reflected field (the interface{}) was
used when attempting to assign elements or associations. This commit
ensures that the typo from the source is used instead in this case.