Make Object constructor accept nested objects in the form of a Hash
This commit ensures that the Object constructor will allow two forms
for a nested Object. The first form is an instance of that Object (like
before) and the second form is a Hash that conforms to the Init type
of that object.

This functionality was previously present in the YAML parser. Placing
it in the Object itself makes i generic and also makes it easy to
create nested instances that contain anonymous types.