Don't include nil values from Go struct when creating init hash
When the initialization hash used when creating an Object instance stems
from a Go struct, it must not include those fields that have a nil
value. The reason for this is that an Object might well declare an
attribute of Boolean type with a default value of false or true. Such
an attribute will require that any value assigned to this attribute must
be of type Boolean. It is not acceptable to assign Undef. To include an
explicit undef value in a initialization hash implies such an illegal
assigment. Such entries must therefore not be included at all. The
Object constructor will then fall back to the default value.