Make it possible to use Hash and Array as keys in a Hash
Prior to this commit, an attempt to use an array as a hash key would
fail with a somewhat cryptic error stating

"invalid argument type for function ToKey, argument 0. expected 'value
 used as hash key', got Tuple[<element types>]".

This commit makes it possible to use both Array and Hash values as keys
in a Hash and corrects the error messages for types that cannot
be used.

Closes lyraproj/lyra#236