Create InstanceHash method on the ObjectType
Several implementations of PuppetObject have a hard-coded InitHash
method. This is unnecessary because a PuppetObject is always an instance
of an ObjectType and the ObjectType has all information needed to create
such a hash dynamically.

This commit adds the ObjectType.InstanceHash() method and replaces the
body of all hard-coded InitHash methods with a call to that method.