Ensure that undef is correctly reflected into map[<key type>]interface{}
This commit ensures that whenever an attempt is made to store an undef
value in a reflected map of type map[<key type>]interface{}, the actual
value is a verbatim `undef` rather than a `interface{}` which is nil.
The reason for this is that the `reflect.Value.SetMapIndex()` method
will otherwise delete the entry. It is a known flaw in the reflect
package that an entry with a `nil` cannot be stored in such a map.

The commit also ensures that the `UndefValue` renders itself correctly
as `null` when serialized into YAML or JSON.

Relates to lyraproj/hiera#37