Make NewGoRuntimeType accept more arguments.
The NewGoRuntimeType previously only accepted an interface{} value that
it did a `reflect.TypeOf()` call on to obtain the actual type. This
commit enables that a `reflect.Type` or a `reflect.Value` can be passed
as an argument. The former will be used as is while the second will be
queried for its type.